aptly-latest-snapshots: Add some basic tests
I was slightly horrified that past me did not write tests for this.
Unfortunately, aptly-rest-mock doesn't implement the publish API...but
at least adding some very basic tests w/ axum should help to make sure
nothing really exploded across the upgrade.
