+++ aptly-latest-snapshots/Cargo.toml
@@ -19,3 +19,7 @@ tracing = "0.1.41"
 tracing-error = "0.2.1"
 tracing-subscriber = "0.3.19"
 url = "2.5.4"
+
+[dev-dependencies]
+axum-test = { version = "17.3.0", features = ["typed-routing"] }
+rstest = "0.25.0"
